Added Store class to store all the information about products
Store is a singleton class (so that all classes can use a single instance of the Store). It keeps product information(name/ price/ category) in data structures(HashMap). At "getInstance()" method it creates the data structures containing name, price and category. It keeps separate HashMaps which maps the name/price/ category of a product with its serial number. Serial numbers are defined in the ProductConst class.

The main method gets instance of the Store class at the beginning.